Updated some tests to use renamed method, having issues with getting location due to location not updating fast enough (GPS updating slowly and unable to test Network at the moment)
