+++ app/src/androidTest/java/com/example/bespinaf/a2d2/test_driver_drops_off_rider.java
@@ -48,7 +48,7 @@ public class test_driver_drops_off_rider {
     @Before
     public void setUp() {
         if (DataSourceUtils.getCurrentRequests().isEmpty()) {
-            DataSourceUtils.initDateFormatters();
+            DataSourceUtils.initializeDateFormatters();
             try {
                 //Required to load data before trying to perform actions/load page
                 Thread.sleep(5000);
