+++ app/src/androidTest/java/com/example/bespinaf/a2d2/test_driver_takes_job.java
@@ -39,7 +39,7 @@ public class test_driver_takes_job {
 
     @Before
     public void setUp() {
-        DataSourceUtils.initDateFormatters();
+        DataSourceUtils.initializeDateFormatters();
         request = buildRideRequest();
 
         Intent intent = new Intent();
