+++ app/src/main/java/com/example/bespinaf/a2d2/utilities/DataSourceUtils.java
@@ -143,6 +143,11 @@ public class DataSourceUtils {
     }
 
 
+    public static HashMap<String, Request> getCurrentRequests(){
+        return requests;
+    }
+
+
     public static HashMap<String, Request> getRequestsWithStatus(String targetStatus){
         HashMap<String, Request> filteredRequests = new HashMap<>();
